Move assumption-related methods to separate source file
